store commandline options for branch predictors and rename `BranchUnit` to `SkylakeBranchUnit`
