+++ CMakeLists.txt
@@ -98,7 +98,7 @@ set(_CUSTOMHW_SOURCE_FILES
     CustomHWUnits/CustomSourceMgr.cpp
     CustomHWUnits/MCADLSUnit.cpp
     CustomHWUnits/NaiveBranchPredictorUnit.cpp
-    CustomHWUnits/BranchUnit.cpp
+    CustomHWUnits/SkylakeBranchUnit.cpp
    )
 
 set(_CUSTOM_STAGES_SOURCE_FILES
